requirements.txt
apscheduler==3.11.0
fastapi==0.115.6
jose==1.0.0
passlib==1.7.4
pydantic==1.10.12
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.19
uvicorn==0.34.0
